Convert test/test-parse-lint.js to Mocha
